// pkgs/models/src/revisions/types.api.d.ts
import type { Pagination, QuerystringOrgProject, Res, PartialUndefined } from '@specfy/core';
import type { AnalyserJson } from '@specfy/stack-analyser';
import type { DBBlobBase, DBBlobComponentBase, DBBlobDocumentBase, DBBlobProjectBase } from '../blobs/types.js';
import type { BlockLevelZero } from '../documents/types.prosemirror.js';
import type { ApiReview } from '../reviews/types.api.js';
import type { ApiUser } from '../users/types.api.js';
import type { DBRevision } from './types.js';
export type ApiRevision = DBRevision & {
    authors: ApiUser[];
    url: string;
};
export interface ParamsRevision {
    revision_id: string;
}
export type ListRevisions = Res<{
    Querystring: {
        org_id: string;
        project_id: string;
        status?: ApiRevision['status'] | 'all' | 'merged' | 'opened';
        search?: string;
    };
    Success: {
        data: ApiRevision[];
        pagination: Pagination;
    };
}>;
export type ApiBlobCreate = Omit<DBBlobBase, 'createdAt' | 'id' | 'updatedAt'> & (DBBlobComponentBase | DBBlobDocumentBase | DBBlobProjectBase);
export type ApiBlobCreateDocument = DBBlobDocumentBase & Omit<DBBlobBase, 'createdAt' | 'id' | 'updatedAt'>;
export type ApiBlobCreateComponent = DBBlobComponentBase & Omit<DBBlobBase, 'createdAt' | 'id' | 'updatedAt'>;
export type ApiBlobCreateProject = DBBlobProjectBase & Omit<DBBlobBase, 'createdAt' | 'id' | 'updatedAt'>;
export type PostRevision = Res<{
    Body: Pick<ApiRevision, 'description' | 'name' | 'orgId' | 'projectId'> & {
        blobs: ApiBlobCreate[];
        draft: boolean;
    };
    Success: {
        data: Pick<ApiRevision, 'id'>;
    };
}>;
export type CreateRevisionError = {
    error: {
        code: 'cant_create';
        reason: 'no_diff';
    };
};
export type PostUploadRevision = Res<{
    Body: Pick<ApiRevision, 'description' | 'name' | 'orgId' | 'projectId'> & {
        source: string;
        blobs: Array<{
            path: string;
            content: string;
        }>;
        stack: AnalyserJson | null;
        autoLayout?: boolean;
    };
    Error: CreateRevisionError;
    Success: {
        data: Pick<ApiRevision, 'id'>;
    };
}>;
export type GetRevision = Res<{
    Params: ParamsRevision;
    Querystring: QuerystringOrgProject;
    Success: {
        data: ApiRevision & {
            reviewers: ApiUser[];
        };
    };
}>;
export type PatchRevision = Res<{
    Params: ParamsRevision;
    Querystring: QuerystringOrgProject;
    Body: PartialUndefined<Pick<GetRevision['Success']['data'], 'description' | 'locked' | 'name' | 'status'> & {
        authors: string[];
        reviewers: string[];
    }>;
    Success: {
        data: {
            done: boolean;
        };
    };
}>;
export type MergeRevisionError = {
    error: {
        code: 'cant_merge';
        reason: 'already_merged' | 'empty' | 'no_reviews' | 'outdated';
    };
};
export type MergeRevision = Res<{
    Params: ParamsRevision;
    Querystring: QuerystringOrgProject;
    Error: MergeRevisionError;
    Success: {
        data: {
            done: true;
        };
    };
}>;
export type CommentRevision = Res<{
    Params: ParamsRevision;
    Querystring: QuerystringOrgProject;
    Body: {
        content: BlockLevelZero;
        approval: boolean;
    };
    Success: {
        data: {
            id: string;
        };
    };
}>;
export type ListRevisionChecks = Res<{
    Params: ParamsRevision;
    Querystring: QuerystringOrgProject;
    Success: {
        data: {
            canMerge: boolean;
            reviews: ApiReview[];
            outdatedBlobs: string[];
        };
    };
}>;
export type RebaseRevision = Res<{
    Params: ParamsRevision;
    Querystring: QuerystringOrgProject;
    Success: {
        data: {
            done: boolean;
        };
    };
}>;
export type DeleteRevision = Res<{
    Params: ParamsRevision;
    Querystring: QuerystringOrgProject;
    Success: never;
}>;